+++ src/Products/ZCatalog/CatalogAwareness.py
@@ -15,7 +15,9 @@
 **NOTE**: This module is deprecated, and should only be used for
   backward-compatibility.  All new code should use CatalogPathAwareness.
 """
+
 import urllib
+import warnings
 
 from Acquisition import aq_base
 from App.special_dtml import DTMLFile
@@ -34,13 +36,18 @@ class CatalogAware:
 
     manage_editCatalogerForm=DTMLFile('dtml/editCatalogerForm', globals())
 
+    def _warn_deprecated(self):
+        warnings.warn('The Products.ZCatalog.CatalogAwareness module is '
+                      'deprecated and will be removed in Zope 2.14. Please '
+                      'use the CatalogPathAwareness module instead.',
+                      DeprecationWarning, stacklevel=3)
+
     def manage_editCataloger(self, default, REQUEST=None):
         """ """
         self.default_catalog=default
         message = "Your changes have been saved"
         if REQUEST is not None:
             return self.manage_main(self, REQUEST, manage_tabs_message=message)
-
 
     def manage_afterAdd(self, item, container):
         self.index_object()
@@ -118,11 +125,13 @@ class CatalogAware:
 
     def index_object(self):
         """A common method to allow Findables to index themselves."""
+        self._warn_deprecated()
         if hasattr(self, self.default_catalog):
             getattr(self, self.default_catalog).catalog_object(self, self.url())
 
     def unindex_object(self):
         """A common method to allow Findables to unindex themselves."""
+        self._warn_deprecated()
         if hasattr(self, self.default_catalog):
             getattr(self, self.default_catalog).uncatalog_object(self.url())